CI: テスト / wasmビルド / 整形 / lint + CPU照合 (docs/ci.md つき) (#22)
* CIを作る (テスト / wasmビルド / web整合) + tools/build-web.sh

## CIの形 — fuurai-dungeon の型

ジョブは「CI」1本に統合し、各ステージの結果は本文つきの Check run として
発行する。ジョブ標準のCheckは選択してもログページへのリンクにしかならず、
リンクだけの項目で Checks タブが埋まるのは意味が無いため。
各ステージは continue-on-error で最後まで走らせ、全レポートを発行してから
まとめて合否判定する。回すのは PR と main へのマージ。

## ステージの選び方 — 手元で回していないものを本命にする

- **wasmビルドが本命。** ネイティブのテストが全部通るのにブラウザでだけ
  壊れる事故 (std::time::Instant は wasm32 に無い等) を止められるのは
  このステージだけ。成果物のサイズも表で出す
- テストは保険 (手元で毎回回している)
- web整合は「糊だけ新しい」事故の検査 (下記)

入れていないもの: cosim は Unicorn のビルドが数分で毎PRの値段に合わない。
clippy (32件) / fmt (差分3568行) は今ゲートにすると赤いまま始まり、
警告のみにすると「緑だが指摘あり」という読めない Check になるので置かない。
掃除コミットと同時に次のPRでゲートとして足す。

## tools/build-web.sh

キャッシュ破りの ?v= を手で上げていた。1日で20回近く、4ファイルにわたって。
上げ忘れれば古いコードが走り、片方だけ上げれば emu.key is not a function
になる。手順書に「気をつけて」と書いてある作業はいずれ必ず失敗する。

- wasm を焼いてから ?v= を web/ 全体で1つの番号に揃えて上げる
  (焼く前に上げない — 失敗したのに番号だけ進むのが一番たちが悪い)
- 実際、リポジトリは既に 6 / 11 / 21 とバラバラだった。対象ごとには
  揃っていたので動いていたが、「たまたま動いている」状態だった。22に統一
- --check は揃っているかだけ見る。CIの web整合 ステージがこれを回す
- mapfile も配列も使わない。macOS の bash は 3.2 で mapfile が無い
  (実際に落ちた)。CI の ubuntu と手元の Mac で同じに動く書き方だけを使う

94テスト通過、--check 通過を確認済み。CI自体の動作はこのPRで確かめる。

* cosim を「CPUを触ったときだけ」回すワークフローにする

co-simが捕まえられるのはCPUの意味論の変化だけである。docs や web の
PRで回しても何も検査していないし、Unicorn のCコンパイルに初回数分かかる。
逆にCPUを触ったPRで回さないと、バグを捕まえられる唯一のタイミング
(マージ前) を逃して、マージ後に main が赤くなってから気づくことになる。

paths で core/src/** / cosim/** / Cargo.toml に絞る。合わないPRには
項目ごと現れない — 「スキップ」という読めないCheckも出さない。

Tier 3a (プロテクトモード) 中はほぼ全PRがCPUを触るので実質毎回になるが、
それはCPUを毎回変えているからで、正しい。キャッシュが温まれば数分で済む。

ワークフローファイル自身も paths に入れてあるので、このPRが初回の
被験者になる (コールドキャッシュのUnicornビルド込み)。

* cargo fmt --all を初めて通す (機械的な整形のみ、3568行)

これまで rustfmt を一度も走らせていなかった。CIで fmt --check を
ゲートにする前提の掃除で、手による変更は無い。94テスト通過。

* clippy を0件にする

CIでゲートにする前の掃除。機械的な --fix に加えて:

- kbd_shift / kbd_peeked を削除 — 書くだけで誰も読まない死骸だった
  (キーボードの状態をBDAへ移した名残)。毎ビルド警告が出ていた
- 意図的な書き方には理由つきで allow を付けた:
  - tests/interrupts.rs の `1 * 4` は「ベクタ番号 × 表の1要素4バイト」。
    --fix が `4` に畳んでしまったので**戻して** allow (畳むと読めない)
  - bios.rs scroll_window の引数8個は INT 10h のレジスタ割り当てそのまま

94テスト通過。

* Checkを「CI 結果」1つに濃縮し、意味の分かる言葉にし、整形とlintをゲート化

## 左リストをリンク集にしない

ステージごとに Check run を発行したら、左リストが4項目に膨らんで
リンク集と見分けがつかなくなった。本文つきCheckを「CI 結果」1つに統合し、
先頭に各ステージの合否表、下に全レポートを並べる。見出しは「4/5 合格」の
形式で、開かなくても落ちた数が分かる。

## 内部用語をやめる

- 「web整合」→「JSとwasmの版ずれ」。本文も ?v= を知らない人が読める文にした
- 「cosim」→「CPU照合」。co-sim は内輪の呼び名で、Checkの名前としては
  何の照合か分からない

## 整形と lint をゲートにする

前段の2コミットで fmt (3568行) と clippy (32件) を掃除したので、
「緑だが指摘あり」を作らずにゲートにできる。clippy は -D warnings で
警告もエラー扱い。fmt が落ちたときは「cargo fmt --all で直る」と
レポートに書いてある。

* 版ずれ検査を廃止し (?v= ごと消す)、Checkを結果だけに削り、docs/ci.md を書く

## 検査で見張るより、事故れない構造にする

「?v= が揃っているか」の検査ステージを作った直後に気づいた:
**キャッシュ問題は serve.py がとっくに解決していた**。全応答に
Cache-Control: no-store を送っている。同じ問題を2回別の方法で直し、
手作業の方 (?v=) だけが残っていた。

番号そのものを廃止した。上げ忘れも、ずれも、ずれの検査も、仕組みごと
消えた。build-web.sh は wasm-pack の包み紙に縮んだ。
FreeDOS が A:\> まで到達することをブラウザで確認済み。

将来 GitHub Pages 等キャッシュする配信に載せるときは、デプロイ側が
コンテンツハッシュを付ける。手で番号を上げる方式には戻さない。

## Checkは「何をしたか」と「結果」だけ

説明の文章を毎回のCheckに載せると、読む人は同じ文を何十回も見る。
説明は docs/ci.md に一度だけ書き、レポートは合否表 (コマンド | 結果) と
ログに削った。cosim のレポートも同様。

## docs/ci.md (新規)

CIの全体図 (mermaid)、ステージごとに止める事故の表、入れていないものの
理由、今後の展望。正直に書いた穴: **OS起動テストはCIではイメージが無いので
スキップして緑になる**。「94 passed」のうち約10件は何も検証していない。
展望の筆頭は fetch-images.sh をCIで回してこれを本物にすること。

整形ゲートに自分が落ちていたのも直した (clippy --fix が崩した整形を
そのまま push していた)。

* lint が綺麗すぎて落ちるのを直す

初回の改訂版CIで lint だけが ❌、しかもログが空だった。原因は
ノイズ除去の grep:

  cargo clippy ... | tee | grep -vE 'Checking|Compiling|Finished'

指摘ゼロだと全行がノイズ除去に食われ、grep が「一致なし = exit 1」を
返し、pipefail がそれを拾う。**綺麗なのに落ちる。**

コマンドの終了コードを先にファイル経由で確保し、整形と分離した。
テスト系のステージにも同じ構造があったので揃えて直した (あちらは
「running N tests」が必ず残るので発現していなかっただけ)。

+++ core/tests/bios.rs
@@ -30,9 +30,9 @@ fn machine() -> Machine {
     let mut m = Machine::new();
     m.load_boot_sector(&sector).unwrap();
     m.cpu.regs[4] = 0x7C00; // SP
-    // **割り込みを開けておく。** キーは 8042 → IRQ1 → INT 09h → BDAの待ち行列 →
-    // INT 16h という順で届く。実BIOSと同じ経路にしたので、割り込みを止めたままだと
-    // INT 09h が走らず、待ち行列に一文字も積まれない
+                            // **割り込みを開けておく。** キーは 8042 → IRQ1 → INT 09h → BDAの待ち行列 →
+                            // INT 16h という順で届く。実BIOSと同じ経路にしたので、割り込みを止めたままだと
+                            // INT 09h が走らず、待ち行列に一文字も積まれない
     m.cpu.set_flag(rustx86_core::cpu::IF, true);
     m
 }
@@ -70,7 +70,11 @@ fn int16_translates_scancodes_to_ascii() {
     m.cpu.regs[AX] = 0x0000; // AH=00: 待って1つ取る
     call_int(&mut m, 0x16, 100_000);
     assert_eq!(m.cpu.regs[AX] as u8, b'A', "AL に ASCII");
-    assert_eq!((m.cpu.regs[AX] >> 8) as u8, 0x1E, "AH にスキャンコード (Aの位置)");
+    assert_eq!(
+        (m.cpu.regs[AX] >> 8) as u8,
+        0x1E,
+        "AH にスキャンコード (Aの位置)"
+    );
 }
 
 /// 記号もShiftの上げ下げを解釈して組み立てる
@@ -128,10 +132,10 @@ fn int16_blocks_until_a_key_arrives() {
 #[test]
 fn int16_translates_control_combinations() {
     for (keys, want) in [
-        ("\u{3}", 0x03u8),  // Ctrl+C
-        ("\u{4}", 0x04),    // Ctrl+D
-        ("\u{1a}", 0x1a),   // Ctrl+Z
-        ("\u{1b}", 0x1b),   // Ctrl+[ は Esc と同じ
+        ("\u{3}", 0x03u8), // Ctrl+C
+        ("\u{4}", 0x04),   // Ctrl+D
+        ("\u{1a}", 0x1a),  // Ctrl+Z
+        ("\u{1b}", 0x1b),  // Ctrl+[ は Esc と同じ
     ] {
         let mut m = machine();
         // 8042へは「Ctrlを押す → キーを押す → 離す」の順で流れる
@@ -166,7 +170,11 @@ fn int16_reports_ctrl_in_the_shift_state() {
     }
     m.cpu.regs[AX] = 0x0200;
     call_int(&mut m, 0x16, 10_000);
-    assert_eq!(m.cpu.regs[AX] as u8 & 0x04, 0x04, "Ctrlのビットが立っていない");
+    assert_eq!(
+        m.cpu.regs[AX] as u8 & 0x04,
+        0x04,
+        "Ctrlのビットが立っていない"
+    );
 }
 
 /// AH=01 は取らずに覗く。無ければ ZF=1
@@ -298,7 +306,10 @@ fn crtc_start_address_scrolls_the_screen_without_touching_memory() {
         Some('B'),
         "開始位置を動かしても画面が変わらない"
     );
-    assert!(m.take_vram_dirty(), "画面が変わったのに描き直しの合図が出ていない");
+    assert!(
+        m.take_vram_dirty(),
+        "画面が変わったのに描き直しの合図が出ていない"
+    );
 }
 
 /// AH=09 はカーソル位置に文字と属性を繰り返し置く
@@ -335,7 +346,7 @@ fn int10_writes_a_string() {
         m.write8(SRC + i as u32, *c);
     }
     m.cpu.sregs[0] = 0; // ES
-    m.cpu.regs[5] = SRC as u32; // BP
+    m.cpu.regs[5] = SRC; // BP
     m.cpu.regs[AX] = 0x1300;
     m.cpu.regs[BX] = 0x0A00; // 明るい緑
     m.cpu.regs[CX] = 2;
